+++ app-install.sh
@@ -22,13 +22,13 @@ function show_usage
 
   echo
   echo "Example:"
-  outputComment "  ${0} --user=phpmyadmin --domain=phpmyadmin.php79.net --app=phpmyadmin --php=70"
+  outputComment "  ${0} --user=phpmyadmin --domain=phpmyadmin.php79.net --app=phpmyadmin --php=71 --ssl"
   echo
-  outputComment "  ${0} --user=wordpress --domain=wordpress.php79.net --app=wordpress --php=70"
+  outputComment "  ${0} --user=wordpress --domain=wordpress.php79.net --app=wordpress --php=71"
   echo
-  outputComment "  ${0} --user=laravel55 --domain=laravel55.php79.net --app=laravel55 --php=70"
+  outputComment "  ${0} --user=laravel55 --domain=laravel55.php79.net --app=laravel55 --php=71 --ssl"
   echo
-  outputComment "  ${0} --user=octobercms --domain=octobercms.php79.net --app=laravel51 --php=70 --skip-install"
+  outputComment "  ${0} --user=octobercms --domain=octobercms.php79.net --app=laravel51 --php=71 --ssl --skip-install"
   echo
 
   echo
@@ -53,11 +53,8 @@ function show_usage
   echo -n "  "
   outputInfo  "--app"
   echo "       설치할 프로그램명은 ${STACK_ROOT}/apps 디렉토리안의 하위 디렉토리명과 일치해야 합니다."
-  echo "         laravel55  - Laravel 5.5"
-  echo "         laravel54  - Laravel 5.4"
-  echo "         laravel53  - Laravel 5.3"
-  echo "         laravel52  - Laravel 5.2"
   echo "         laravel51  - Laravel 5.1"
+  echo "         laravel58  - Laravel 5.8  (5.1 ~ 5.8 까지 지원)"
   echo "         wordpress  - WordPress"
   echo "         phpmyadmin - phpMyAdmin"
   echo "         gnuboard5  - 그누보드 5"
@@ -67,7 +64,7 @@ function show_usage
 
   echo -n "  "
   outputInfo  "--php"
-  echo "       PHP 버전을 [ 53 54 55 56 56 70 71 72 ] 형식으로 하나만 입력하세요."
+  echo "       PHP 버전을 [ 53 54 55 56 56 70 71 72 73 ] 형식으로 하나만 입력하세요."
   echo "                Tip) Laravel 은 70, 그누보드4 는 53 등 프로그램에 따라 적절히 선택하세요."
   echo "                     ./status.sh 명령을 통해 현재 서버에 설치된 PHP 버전을 확인할 수 있습니다."
   echo
@@ -76,6 +73,14 @@ function show_usage
   outputInfo  "--skip-install"
   echo "       (선택) 계정 추가 및 nginx 설정까지만 진행하고, 앱 자동 설치(install.sh)는 생략합니다."
   echo "                Tip) 이미 제작된 소스로 설치하거나, 직접 소스를 설치하실 경우에 사용하세요."
+  echo
+
+  echo -n "  "
+  outputInfo  "--ssl"
+  echo "       (선택) Let's Encrypt 자동화툴을 사용해서, SSL 인증서 발급을 함께 진행합니다."
+  echo "                주의) 인증서를 발급 받은 도메인의 IP가 현재 서버로 지정되어 있어야 인증서 발급이 가능합니다."
+  echo "                      앱 설치가 완료된 마지막 단계에서 발급을 시도하므로, 발급 실패시에도 앱 사용에는 문제없습니다."
+  echo "                      앱 설치후에도 ssl-install.sh 명령을 통해 별도 발급 가능합니다."
   echo
 }
 
@@ -92,6 +97,7 @@ if [ -z ${1} ]; then
   exit
 else
   INPUT_SKIP_INSTALL=0
+  INPUT_SSL=0
   for i in "${@}"
   do
     case $i in
@@ -119,6 +125,10 @@ else
       shift
       INPUT_SKIP_INSTALL=1
       ;;
+    --ssl)
+      shift
+      INPUT_SSL=1
+      ;;
     -h | --help )
       show_usage
       exit
@@ -139,22 +149,6 @@ if [ -z ${INPUT_PASSWORD} ]; then
   PASSWORD_GENERATED=1
 fi
 
-# 비밀번호도 입력받아야 하므로, 계정 추가 작업도 일괄 처리
-./user-add.sh --user=${INPUT_USER} --password=${INPUT_PASSWORD} --skip-guide-app-install
-if [ "${?}" != "0" ]; then
-  abort "시스템 계정 추가 작업이 실패하였습니다."
-fi
-echo
-
-#if [ -z $(id -u ${INPUT_USER}) ]; then
-#  input_abort "존재하지 않는 user 입니다.  ./user-add.sh 로 먼저 추가하세요."
-#fi
-
-# TODO: /home 대신 실제 사용자의 HOME_DIR 을 읽어오도록 개선 필요
-if [ ! -d "/home/${INPUT_USER}" ]; then
-  input_abort "/home/${INPUT_USER} 디렉토리가 존재하지 않습니다.   실제 존재하는 시스템 계정인지 확인해주세요."
-fi
-
 if [ -z ${INPUT_DOMAIN} ]; then
   input_abort "domain 항목을 입력하세요."
 fi
@@ -172,7 +166,7 @@ if [ -z ${INPUT_PHP_VERSION} ]; then
 fi
 
 if [ ! -f "/usr/bin/php${INPUT_PHP_VERSION}" ]; then
-  input_abort "PHP ${INPUT_PHP_VERSION} 버전은 아직 설치되지 않았습니다.  입력 형식) 53 54 55 56 70 71 72"
+  input_abort "PHP ${INPUT_PHP_VERSION} 버전은 아직 설치되지 않았습니다.  입력 형식) 53 54 55 56 70 71 72 73"
 fi
 
 # nginx 중복 체크
@@ -200,6 +194,17 @@ if [ "${?}" != "0" ]; then
   abort "nginx 설정에 문제가 있습니다."
 fi
 
+# 비밀번호도 입력받아야 하므로, 계정 추가 작업도 일괄 처리
+./user-add.sh --user=${INPUT_USER} --password=${INPUT_PASSWORD} --skip-guide-app-install
+if [ "${?}" != "0" ]; then
+  abort "시스템 계정 추가 작업이 실패하였습니다."
+fi
+echo
+
+# TODO: /home 대신 실제 사용자의 HOME_DIR 을 읽어오도록 개선 필요
+if [ ! -d "/home/${INPUT_USER}" ]; then
+  input_abort "/home/${INPUT_USER} 디렉토리가 존재하지 않습니다.   실제 존재하는 시스템 계정인지 확인해주세요."
+fi
 
 # nginx 설정 추가
 notice "nginx 에 새로운 사이트 설정을 추가합니다."
@@ -254,6 +259,15 @@ fi
 # 비밀번호 자동 생성시, ~/.my.cnf 생성
 if [ ${PASSWORD_GENERATED} = "1" ]; then
   su - ${INPUT_USER} -c "printf \"[client]\\npassword=${INPUT_PASSWORD}\\n\" > ~/.my.cnf && chmod go-rwx ~/.my.cnf"
+fi
+
+# SSL 인증서 발급 및 nginx SSL 설정 추가 - 실패시에도 설치 완료 화면 표시
+if [ ${INPUT_SSL} = "1" ]; then
+   if [ -f "/usr/bin/certbot-auto" ]; then
+     ${STACK_ROOT}/ssl-install.sh --user=${INPUT_USER} --domain=${INPUT_DOMAIN}
+   else
+     outputError "Let's Encrypt 자동화툴이 설치되지 않았습니다. SSL 인증서 발급을 생략합니다."
+   fi
 fi
 
 echo